migrate ProNewCompanyAccess
