+++ app/controllers/AdminController.scala
@@ -31,6 +31,7 @@ import services.emails.EmailDefinitionsAdmin.AdminProbeTriggered
 import services.emails.EmailDefinitionsDggcrf._
 import services.emails.EmailDefinitionsPro.ProCompaniesAccessesInvitations
 import services.emails.EmailDefinitionsPro.ProCompanyAccessInvitation
+import services.emails.EmailDefinitionsPro.ProNewCompanyAccess
 import services.emails.EmailDefinitionsVarious.ResetPassword
 import services.emails.EmailDefinitionsVarious.UpdateEmailAddress
 import services.emails.EmailsExamplesUtils._
@@ -127,7 +128,8 @@ class AdminController(
     DgccrfReportNotification,
     DgccrfValidateEmail,
     ProCompanyAccessInvitation,
-    ProCompaniesAccessesInvitations
+    ProCompaniesAccessesInvitations,
+    ProNewCompanyAccess
   ).flatMap(readExamplesWithFullKey)
 
   val availableEmails = List[(String, EmailAddress => Email)](
@@ -135,7 +137,6 @@ class AdminController(
 
     // ======= PRO =======
 
-    "pro.new_company_access"     -> (recipient => ProNewCompanyAccess(recipient, genCompany, None)),
     "pro.new_companies_accesses" -> (recipient => ProNewCompaniesAccesses(recipient, genCompanyList, genSiren)),
     "pro.report_ack_pro" -> (recipient =>
       ProResponseAcknowledgment(genReport, genReportResponse, genUser.copy(email = recipient))
